requirements.txt
pygooglevoice>=0.5
python-dotenv>=1.0.0
requests>=2.31.0
